NX-493: Separated out generator from producer
To be able to get properly post-processed packages but not
in "raw" (content locator ) form
